tests/e2e: Refactor gRPC mTLS tests
This is to make it easier to re-use and extend tests later.

Relates: ff175ad13437621b8a4d6250e3bcd75e0ba52dbd
